Added key checking, patched bugs, & added omit desync

+++ src/commands/default.rs
@@ -5,10 +5,6 @@ use crate::components::{self, linker::{read_linker, write_linker, LinkedObject, 
 pub fn run_default(subcommand: &str) {
     let mut path_buf = PathBuf::from(subcommand);
 
-    if path_buf.is_relative() {
-        path_buf = env::current_dir().unwrap().join(path_buf);
-    }
-
     match path_buf.try_exists() {
         Ok(exists) => {
 
@@ -47,21 +43,40 @@ pub fn run_default(subcommand: &str) {
 
                         None => {
 
+                            let found_existant = linked_objects.iter().find(|linked_object| linked_object.path == path_str);
+
+                            if found_existant.is_some() {
+                                let encrypted_path = format!(".omit/{}", found_existant.unwrap().identifier);
+                                let encrypted_path_buf = PathBuf::from(encrypted_path);
+
+                                fs::write(&encrypted_path_buf, encrypted_content).expect("Unable to write encrypted file");
+                                return;
+                            }
+
                             let generated_identifier = cuid::cuid2();
 
                             let encrypted_path = format!(".omit/{}", generated_identifier);
                             let encrypted_path_buf = PathBuf::from(encrypted_path);
                             fs::write(encrypted_path_buf, &encrypted_content).expect("Unable to write encrypted file");
 
-                            linked_objects.push(components::linker::LinkedObject::new(generated_identifier, path_str.to_string(), encrypted_content[0..9].to_string()));
-                            write_linker(LinkerPage { linked_objects }).expect("Unable to write linker file");
+                            let to_strip = std::env::current_dir().unwrap();
+                            let strip_str = to_strip.to_str().unwrap();
+
+                            if path_buf.is_relative() {
+                                linked_objects.push(components::linker::LinkedObject::new(generated_identifier, path_buf.to_str().unwrap().to_string(), encrypted_content[0..9].to_string()));
+                            } else {
+                                linked_objects.push(components::linker::LinkedObject::new(generated_identifier, path_buf.strip_prefix(strip_str).unwrap().to_str().unwrap().to_string(), encrypted_content[0..9].to_string()));
+                            }
+                            
+                            write_linker(LinkerPage { 
+                                linked_objects: linked_objects,
+                                key_check: linker_page.key_check
+                            }).expect("Unable to write linker file");
                         }
                     }
 
                     println!("Ensuring all secrets...");
                     crate::commands::ensure::run_ensure();
-
-                    
                 },
                 Err(_) => {
                     eprintln!("Couldn't find key, run 'omit key [key]' to generate a key");
